[windows] AE: refactor windows sinks to avoid duplicated code.
