+++ xbmc/cores/AudioEngine/AESinkFactory.cpp
@@ -20,12 +20,13 @@
 
 #include "AESinkFactory.h"
 #include "Interfaces/AESink.h"
-#if defined(TARGET_WINDOWS_DESKTOP)
+#if defined(TARGET_WINDOWS)
   #include "Sinks/AESinkWASAPI.h"
-  #include "Sinks/AESinkDirectSound.h"
-#elif defined(TARGET_WINDOWS_STORE)
-  #include "Sinks/AESinkXAudioWin10.h"
-  #include "Sinks/AESinkWASAPIWin10.h"
+  #if defined(TARGET_WINDOWS_DESKTOP)
+    #include "Sinks/AESinkDirectSound.h"
+  #else // TARGET_WINDOWS_STORE
+    #include "Sinks/AESinkXAudio.h"
+  #endif // TARGET_WINDOWS
 #elif defined(TARGET_ANDROID)
   #include "Sinks/AESinkAUDIOTRACK.h"
 #elif defined(TARGET_RASPBERRY_PI)
@@ -68,12 +69,13 @@ void CAESinkFactory::ParseDevice(std::string &device, std::string &driver)
 
     // check that it is a valid driver name
     if (
-#if defined(TARGET_WINDOWS_DESKTOP)
+#if defined(TARGET_WINDOWS)
         driver == "WASAPI"      ||
+  #if defined(TARGET_WINDOWS_DESKTOP)
         driver == "DIRECTSOUND" ||
-#elif defined(TARGET_WINDOWS_STORE)
+  #else // TARGET_WINDOWS_STORE
         driver == "XAUDIO"      ||
-        driver == "WASAPI"      ||
+  #endif
 #elif defined(TARGET_ANDROID)
         driver == "AUDIOTRACK"  ||
 #elif defined(TARGET_RASPBERRY_PI)
@@ -116,16 +118,16 @@ IAESink *CAESinkFactory::TrySink(const std::string &driver, std::string &device,
     sink = new CAESinkNULL();
   else
   {
-#if defined(TARGET_WINDOWS_DESKTOP)
+#if defined(TARGET_WINDOWS)
     if (driver == "WASAPI")
       sink = new CAESinkWASAPI();
+  #if defined(TARGET_WINDOWS_DESKTOP)
     else if (driver == "DIRECTSOUND")
       sink = new CAESinkDirectSound();
-#elif defined(TARGET_WINDOWS_STORE)
-    if (driver == "WASAPI")
-      sink = new CAESinkWASAPIWin10();
+  #else // TARGET_WINDOWS_STORE
     else if (driver == "XAUDIO")
-      sink = new CAESinkXAudioWin10();
+      sink = new CAESinkXAudio();
+  #endif
 #elif defined(TARGET_ANDROID)
     sink = new CAESinkAUDIOTRACK();
 #elif defined(TARGET_RASPBERRY_PI)
@@ -202,7 +204,8 @@ IAESink *CAESinkFactory::Create(std::string &device, AEAudioFormat &desiredForma
 void CAESinkFactory::EnumerateEx(AESinkInfoList &list, bool force)
 {
   AESinkInfo info;
-#if defined(TARGET_WINDOWS_DESKTOP)
+#if defined(TARGET_WINDOWS)
+  #if defined(TARGET_WINDOWS_DESKTOP)
 
   info.m_deviceInfoList.clear();
   info.m_sinkName = "DIRECTSOUND";
@@ -210,25 +213,22 @@ void CAESinkFactory::EnumerateEx(AESinkInfoList &list, bool force)
   if(!info.m_deviceInfoList.empty())
     list.push_back(info);
 
+  #else // TARGET_WINDOWS_STORE
+
+  info.m_deviceInfoList.clear();
+  info.m_sinkName = "XAUDIO";
+  CAESinkXAudio::EnumerateDevicesEx(info.m_deviceInfoList, force);
+  if (!info.m_deviceInfoList.empty())
+    list.push_back(info);
+
+  #endif
+
   info.m_deviceInfoList.clear();
   info.m_sinkName = "WASAPI";
   CAESinkWASAPI::EnumerateDevicesEx(info.m_deviceInfoList, force);
   if(!info.m_deviceInfoList.empty())
     list.push_back(info);
 
-#elif defined(TARGET_WINDOWS_STORE)
-
-  info.m_deviceInfoList.clear();
-  info.m_sinkName = "XAUDIO";
-  CAESinkXAudioWin10::EnumerateDevicesEx(info.m_deviceInfoList, force);
-  if (!info.m_deviceInfoList.empty())
-    list.push_back(info);
-
-  info.m_deviceInfoList.clear();
-  info.m_sinkName = "WASAPI";
-  CAESinkWASAPIWin10::EnumerateDevicesEx(info.m_deviceInfoList, force);
-  if (!info.m_deviceInfoList.empty())
-    list.push_back(info);
 #elif defined(TARGET_ANDROID)
 
   info.m_deviceInfoList.clear();
