+++ xbmc/cores/AudioEngine/CMakeLists.txt
@@ -86,16 +86,20 @@ if(FFMPEG_FOUND)
   list(APPEND HEADERS Engines/ActiveAE/ActiveAEResampleFFMPEG.h)
 endif()
 
-if(CORE_SYSTEM_NAME STREQUAL windowsstore)
-  list(APPEND SOURCES Sinks/AESinkWASAPIWin10.cpp
-                      Sinks/AESinkXAudioWin10.cpp)
-  list(APPEND HEADERS Sinks/AESinkWASAPIWin10.h
-                      Sinks/AESinkXAudioWin10.h)
-elseif(CORE_SYSTEM_NAME STREQUAL windows)
-  list(APPEND SOURCES Sinks/AESinkDirectSound.cpp
-                      Sinks/AESinkWASAPI.cpp)
-  list(APPEND HEADERS Sinks/AESinkDirectSound.h
-                      Sinks/AESinkWASAPI.h)
+if(CORE_SYSTEM_NAME MATCHES windows)
+  list(APPEND SOURCES Sinks/AESinkWASAPI.cpp
+                      Sinks/windows/AESinkFactoryWin.cpp)
+  list(APPEND HEADERS Sinks/AESinkWASAPI.h
+                      Sinks/windows/AESinkFactoryWin.h)
+  if(CORE_SYSTEM_NAME STREQUAL windowsstore)
+    list(APPEND SOURCES Sinks/AESinkXAudio.cpp
+                        Sinks/windows/AESinkFactoryWin10.cpp)
+    list(APPEND SOURCES Sinks/AESinkXAudio.h)
+  elseif(CORE_SYSTEM_NAME STREQUAL windows)
+    list(APPEND SOURCES Sinks/AESinkDirectSound.cpp
+                        Sinks/windows/AESinkFactoryWin32.cpp)
+    list(APPEND SOURCES Sinks/AESinkDirectSound.h)
+  endif()
 endif()
 
 if(CORE_PLATFORM_NAME_LC STREQUAL rbpi)
