+++ xbmc/cores/AudioEngine/Sinks/AESinkWASAPI.cpp
@@ -30,154 +30,17 @@
 #include <Mmreg.h>
 #include "utils/StringUtils.h"
 
-#pragma comment(lib, "Avrt.lib")
+#ifdef TARGET_WINDOWS_DESKTOP
+#  pragma comment(lib, "Avrt.lib")
+#endif // TARGET_WINDOWS_DESKTOP
 
-const CLSID CLSID_MMDeviceEnumerator = __uuidof(MMDeviceEnumerator);
-const IID IID_IMMDeviceEnumerator = __uuidof(IMMDeviceEnumerator);
-const IID IID_IAudioClient = __uuidof(IAudioClient);
 const IID IID_IAudioRenderClient = __uuidof(IAudioRenderClient);
 const IID IID_IAudioClock = __uuidof(IAudioClock);
-
-static const unsigned int WASAPISampleRateCount = 10;
-static const unsigned int WASAPISampleRates[] = {384000, 192000, 176400, 96000, 88200, 48000, 44100, 32000, 22050, 11025};
-
-#define WASAPI_SPEAKER_COUNT 21
-static const unsigned int WASAPIChannelOrder[] = {AE_CH_RAW,
-                                                  SPEAKER_FRONT_LEFT,           SPEAKER_FRONT_RIGHT,           SPEAKER_FRONT_CENTER,
-                                                  SPEAKER_LOW_FREQUENCY,        SPEAKER_BACK_LEFT,             SPEAKER_BACK_RIGHT,
-                                                  SPEAKER_FRONT_LEFT_OF_CENTER, SPEAKER_FRONT_RIGHT_OF_CENTER,
-                                                  SPEAKER_BACK_CENTER,          SPEAKER_SIDE_LEFT,             SPEAKER_SIDE_RIGHT,
-                                                  SPEAKER_TOP_FRONT_LEFT,       SPEAKER_TOP_FRONT_RIGHT,       SPEAKER_TOP_FRONT_CENTER,
-                                                  SPEAKER_TOP_CENTER,           SPEAKER_TOP_BACK_LEFT,         SPEAKER_TOP_BACK_RIGHT,
-                                                  SPEAKER_TOP_BACK_CENTER,      SPEAKER_RESERVED,              SPEAKER_RESERVED};
-
-static const enum AEChannel AEChannelNames[]   = {AE_CH_RAW,
-                                                  AE_CH_FL,                     AE_CH_FR,                      AE_CH_FC,
-                                                  AE_CH_LFE,                    AE_CH_BL,                      AE_CH_BR,
-                                                  AE_CH_FLOC,                   AE_CH_FROC,
-                                                  AE_CH_BC,                     AE_CH_SL,                      AE_CH_SR,
-                                                  AE_CH_TFL,                    AE_CH_TFR,                     AE_CH_TFC ,
-                                                  AE_CH_TC  ,                   AE_CH_TBL,                     AE_CH_TBR,
-                                                  AE_CH_TBC,                    AE_CH_BLOC,                    AE_CH_BROC};
-
-static const enum AEChannel layoutsList[][16] = 
-{
-  /* Most common configurations */
-  {AE_CH_FC,  AE_CH_NULL}, // Mono
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_NULL}, // Stereo
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_BL,  AE_CH_BR,  AE_CH_NULL}, // Quad
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BC,  AE_CH_NULL}, // Surround
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_NULL}, // Standard 5.1
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_BL,  AE_CH_BR,  AE_CH_NULL}, // 5.1 wide (obsolete)
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_NULL}, // Standard 7.1
-  /* Less common configurations */
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_LFE, AE_CH_NULL}, // 2.1
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_BL,  AE_CH_BR,  AE_CH_FLOC,AE_CH_FROC,AE_CH_NULL}, // 7.1 wide (obsolete)
-  /* Exotic configurations */
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_NULL}, // 3 front speakers
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_NULL}, // 3 front speakers + LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_BL,  AE_CH_BR,  AE_CH_LFE, AE_CH_NULL}, // Quad + LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BC,  AE_CH_LFE, AE_CH_NULL}, // Surround + LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_SL,  AE_CH_SR,  AE_CH_NULL}, // Standard 5.1 w/o LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BL,  AE_CH_BR,  AE_CH_NULL}, // 5.1 wide w/o LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_SL,  AE_CH_SR,  AE_CH_BC,  AE_CH_NULL}, // Standard 5.1 w/o LFE + Back Center
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BL,  AE_CH_BC,  AE_CH_BR,  AE_CH_NULL}, // 5.1 wide w/o LFE + Back Center
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_BL,  AE_CH_BR,  AE_CH_TC,  AE_CH_NULL}, // DVD speakers
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BL,  AE_CH_BR,  AE_CH_BC,  AE_CH_LFE, AE_CH_NULL}, // 5.1 wide + Back Center
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_NULL}, // Standard 7.1 w/o LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_BL,  AE_CH_BR,  AE_CH_FLOC,AE_CH_FROC,AE_CH_NULL}, // 7.1 wide w/o LFE
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BC,  AE_CH_BR,  AE_CH_NULL}, // Standard 7.1 + Back Center
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_FLOC,AE_CH_FROC,AE_CH_NULL}, // Standard 7.1 + front wide
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_NULL}, // Standard 7.1 + 2 front top
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_TFC, AE_CH_NULL}, // Standard 7.1 + 3 front top
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_TBL, AE_CH_TBR, AE_CH_NULL}, // Standard 7.1 + 2 front top + 2 back top
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_TFC, AE_CH_TBL, AE_CH_TBR, AE_CH_NULL}, // Standard 7.1 + 3 front top + 2 back top
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_TFC, AE_CH_TBL, AE_CH_TBR, AE_CH_TBC, AE_CH_NULL}, // Standard 7.1 + 3 front top + 3 back top
-  {AE_CH_FL,  AE_CH_FR,  AE_CH_FC,  AE_CH_LFE, AE_CH_SL,  AE_CH_SR,  AE_CH_BL,  AE_CH_BR,  AE_CH_TFL, AE_CH_TFR, AE_CH_TFC, AE_CH_TBL, AE_CH_TBR, AE_CH_TBC, AE_CH_TC,  AE_CH_NULL} // Standard 7.1 + 3 front top + 3 back top + Top Center
-};
-
-struct sampleFormat
-{
-  GUID subFormat;
-  unsigned int bitsPerSample;
-  unsigned int validBitsPerSample;
-  AEDataFormat subFormatType;
-};
-
-//! @todo
-//! Sample formats go from float -> 32 bit int -> 24 bit int (packed in 32) -> -> 24 bit int -> 16 bit int */
-//! versions of Kodi before 14.0 had a bug which made S24NE4MSB the first format selected
-//! this bug worked around some driver bug of some IEC958 devices which report S32 but can't handle it
-//! correctly. So far I have never seen and WASAPI device using S32 and don't think probing S24 before
-//! S32 has any negative impact.
-static const sampleFormat testFormats[] = { {KSDATAFORMAT_SUBTYPE_IEEE_FLOAT, 32, 32, AE_FMT_FLOAT},
-                                            {KSDATAFORMAT_SUBTYPE_PCM, 32, 24, AE_FMT_S24NE4MSB},
-                                            {KSDATAFORMAT_SUBTYPE_PCM, 32, 32, AE_FMT_S32NE},
-                                            {KSDATAFORMAT_SUBTYPE_PCM, 24, 24, AE_FMT_S24NE3},
-                                            {KSDATAFORMAT_SUBTYPE_PCM, 16, 16, AE_FMT_S16NE} };
-
-struct winEndpointsToAEDeviceType
-{
-  std::string winEndpointType;
-  AEDeviceType aeDeviceType;
-};
-
-static const winEndpointsToAEDeviceType winEndpoints[EndpointFormFactor_enum_count] =
-{
-  {"Network Device - ",         AE_DEVTYPE_PCM},
-  {"Speakers - ",               AE_DEVTYPE_PCM},
-  {"LineLevel - ",              AE_DEVTYPE_PCM},
-  {"Headphones - ",             AE_DEVTYPE_PCM},
-  {"Microphone - ",             AE_DEVTYPE_PCM},
-  {"Headset - ",                AE_DEVTYPE_PCM},
-  {"Handset - ",                AE_DEVTYPE_PCM},
-  {"Digital Passthrough - ", AE_DEVTYPE_IEC958},
-  {"SPDIF - ",               AE_DEVTYPE_IEC958},
-  {"HDMI - ",                  AE_DEVTYPE_HDMI},
-  {"Unknown - ",                AE_DEVTYPE_PCM},
-};
-
-AEDeviceInfoList DeviceInfoList;
-
-#define EXIT_ON_FAILURE(hr, reason, ...) if(FAILED(hr)) {CLog::Log(LOGERROR, reason " - %s", __VA_ARGS__, WASAPIErrToStr(hr)); goto failed;}
-
-#define ERRTOSTR(err) case err: return #err
-
 DEFINE_PROPERTYKEY(PKEY_Device_FriendlyName, 0xa45c254e, 0xdf1c, 0x4efd, 0x80, 0x20, 0x67, 0xd1, 0x46, 0xa8, 0x50, 0xe0, 14);
 DEFINE_PROPERTYKEY(PKEY_Device_EnumeratorName, 0xa45c254e, 0xdf1c, 0x4efd, 0x80, 0x20, 0x67, 0xd1, 0x46, 0xa8, 0x50, 0xe0, 24);
 
-DWORD ChLayoutToChMask(const enum AEChannel * layout, unsigned int * numberOfChannels = NULL)
-{
-  if (numberOfChannels)
-    *numberOfChannels = 0;
-  if (!layout)
-    return 0;
-  
-  DWORD mask = 0;
-  unsigned int i;
-  for (i = 0; layout[i] != AE_CH_NULL; i++)
-    mask |= WASAPIChannelOrder[layout[i]];
-  
-  if (numberOfChannels)
-    *numberOfChannels = i;
-
-  return mask;
-}
-
-std::string localWideToUtf(LPCWSTR wstr)
-{
-  if (wstr == NULL)
-    return "";
-  int bufSize = WideCharToMultiByte(CP_UTF8, 0, wstr, -1, NULL, 0, NULL, NULL);
-  char *multiStr = new char[bufSize + 1];
-  if (bufSize == 0 || WideCharToMultiByte(CP_UTF8, 0, wstr, -1, multiStr, bufSize, NULL, NULL) != bufSize)
-    multiStr[0] = 0;
-  else
-    multiStr[bufSize] = 0;
-  std::string ret(multiStr);
-  delete[] multiStr;
-  return ret;
-}
+extern const char *WASAPIErrToStr(HRESULT err);
+#define EXIT_ON_FAILURE(hr, reason, ...) if(FAILED(hr)) {CLog::Log(LOGERROR, reason " - %s", __VA_ARGS__, WASAPIErrToStr(hr)); goto failed;}
 
 CAESinkWASAPI::CAESinkWASAPI() :
   m_needDataEvent(0),
@@ -216,87 +79,41 @@ bool CAESinkWASAPI::Initialize(AEAudioFormat &format, std::string &device)
 
   m_device = device;
   bool bdefault = false;
+  HRESULT hr = S_FALSE;
 
   /* Save requested format */
   /* Clear returned format */
   sinkReqFormat = format.m_dataFormat;
   sinkRetFormat = AE_FMT_INVALID;
 
-  IMMDeviceEnumerator* pEnumerator = NULL;
-  IMMDeviceCollection* pEnumDevices = NULL;
-
-  HRESULT hr = CoCreateInstance(CLSID_MMDeviceEnumerator, NULL, CLSCTX_ALL, IID_IMMDeviceEnumerator, (void**)&pEnumerator);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Could not allocate WASAPI device enumerator. CoCreateInstance error code: %li", hr)
-
-  /* Get our device. First try to find the named device. */
-  UINT uiCount = 0;
-
-  hr = pEnumerator->EnumAudioEndpoints(eRender, DEVICE_STATE_ACTIVE, &pEnumDevices);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of audio endpoint enumeration failed.")
-
-  hr = pEnumDevices->GetCount(&uiCount);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of audio endpoint count failed.")
-
   if(StringUtils::EndsWithNoCase(device, std::string("default")))
     bdefault = true;
 
   if(!bdefault)
   {
-    for (UINT i = 0; i < uiCount; i++)
-    {
-      IPropertyStore *pProperty = NULL;
-      PROPVARIANT varName;
-
-      hr = pEnumDevices->Item(i, &m_pDevice);
-      EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of WASAPI endpoint failed.")
-
-      hr = m_pDevice->OpenPropertyStore(STGM_READ, &pProperty);
-      EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of WASAPI endpoint properties failed.")
-
-      hr = pProperty->GetValue(PKEY_AudioEndpoint_GUID, &varName);
-      if (FAILED(hr))
-      {
-        CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint GUID failed.");
-        SAFE_RELEASE(pProperty);
-        goto failed;
-      }
-
-      std::string strDevName = localWideToUtf(varName.pwszVal);
-
-      if (device == strDevName)
-        i = uiCount;
-      else
-        SAFE_RELEASE(m_pDevice);
-
-      PropVariantClear(&varName);
-      SAFE_RELEASE(pProperty);
-    }
+    hr = CAESinkFactoryWin::ActivateWASAPIDevice(device, &m_pDevice);
+    EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of WASAPI endpoint failed.")
   }
-  SAFE_RELEASE(pEnumDevices);
 
   if (!m_pDevice)
   {
     if(!bdefault)
       CLog::Log(LOGINFO, __FUNCTION__": Could not locate the device named \"%s\" in the list of WASAPI endpoint devices.  Trying the default device...", device.c_str());
-    hr = pEnumerator->GetDefaultAudioEndpoint(eRender, eConsole, &m_pDevice);
+
+    std::string defaultId = CAESinkFactoryWin::GetDefaultDeviceId();
+    if (defaultId.empty())
+    {
+      CLog::Log(LOGINFO, __FUNCTION__": Could not locate the default device id in the list of WASAPI endpoint devices.");
+      goto failed;
+    }
+
+    hr = CAESinkFactoryWin::ActivateWASAPIDevice(defaultId, &m_pDevice);
     EXIT_ON_FAILURE(hr, __FUNCTION__": Could not retrieve the default WASAPI audio endpoint.")
 
-    IPropertyStore *pProperty = NULL;
-    PROPVARIANT varName;
-
-    hr = m_pDevice->OpenPropertyStore(STGM_READ, &pProperty);
-    EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of WASAPI endpoint properties failed.")
-
-    hr = pProperty->GetValue(PKEY_AudioEndpoint_GUID, &varName);
-
-    device = localWideToUtf(varName.pwszVal);
-    PropVariantClear(&varName);
-    SAFE_RELEASE(pProperty);
+    device = defaultId;
   }
 
-  SAFE_RELEASE(pEnumerator);
-
-  hr = m_pDevice->Activate(IID_IAudioClient, CLSCTX_ALL, NULL, (void**)&m_pAudioClient);
+  hr = m_pDevice->Activate(&m_pAudioClient);
   EXIT_ON_FAILURE(hr, __FUNCTION__": Activating the WASAPI endpoint device failed.")
 
   if (!InitializeExclusive(format))
@@ -340,8 +157,6 @@ bool CAESinkWASAPI::Initialize(AEAudioFormat &format, std::string &device)
 
 failed:
   CLog::Log(LOGERROR, __FUNCTION__": WASAPI initialization failed.");
-  SAFE_RELEASE(pEnumDevices);
-  SAFE_RELEASE(pEnumerator);
   SAFE_RELEASE(m_pRenderClient);
   SAFE_RELEASE(m_pAudioClient);
   SAFE_RELEASE(m_pAudioClock);
@@ -552,126 +367,36 @@ unsigned int CAESinkWASAPI::AddPackets(uint8_t **data, unsigned int frames, unsi
 
 void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool force)
 {
-  IMMDeviceEnumerator* pEnumerator = NULL;
-  IMMDeviceCollection* pEnumDevices = NULL;
-  IMMDevice*           pDefaultDevice = NULL;
   CAEDeviceInfo        deviceInfo;
   CAEChannelInfo       deviceChannels;
-  LPWSTR               pwszID = NULL;
-  std::wstring         wstrDDID;
   bool                 add192 = false;
 
   WAVEFORMATEXTENSIBLE wfxex = {0};
   HRESULT              hr;
 
-  hr = CoCreateInstance(CLSID_MMDeviceEnumerator, NULL, CLSCTX_ALL, IID_IMMDeviceEnumerator, (void**)&pEnumerator);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Could not allocate WASAPI device enumerator. CoCreateInstance error code: %li", hr)
-
-  UINT uiCount = 0;
-
-  // get the default audio endpoint
-  if(pEnumerator->GetDefaultAudioEndpoint(eRender, eConsole, &pDefaultDevice) == S_OK)
+  for (RendererDetail details : CAESinkFactoryWin::GetRendererDetails())
   {
-    if(pDefaultDevice->GetId(&pwszID) == S_OK)
-    {
-      wstrDDID = pwszID;
-      CoTaskMemFree(pwszID);
-    }
-    SAFE_RELEASE(pDefaultDevice);
-  }
-
-  // enumerate over all audio endpoints
-  hr = pEnumerator->EnumAudioEndpoints(eRender, DEVICE_STATE_ACTIVE, &pEnumDevices);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of audio endpoint enumeration failed.")
-
-  hr = pEnumDevices->GetCount(&uiCount);
-  EXIT_ON_FAILURE(hr, __FUNCTION__": Retrieval of audio endpoint count failed.")
-
-  for (UINT i = 0; i < uiCount; i++)
-  {
-    IMMDevice *pDevice = NULL;
-    IPropertyStore *pProperty = NULL;
-    PROPVARIANT varName;
-    PropVariantInit(&varName);
-
     deviceInfo.m_channels.Reset();
     deviceInfo.m_dataFormats.clear();
     deviceInfo.m_sampleRates.clear();
+    deviceChannels.Reset();
 
-    hr = pEnumDevices->Item(i, &pDevice);
+    for (unsigned int c = 0; c < WASAPI_SPEAKER_COUNT; c++)
+    {
+      if (details.uiChannelMask & WASAPIChannelOrder[c])
+        deviceChannels += AEChannelNames[c];
+    }
+
+    IAEWASAPIDevice* pDevice;
+    hr = CAESinkFactoryWin::ActivateWASAPIDevice(details.strDeviceId, &pDevice);
     if (FAILED(hr))
     {
       CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint failed.");
       goto failed;
     }
 
-    hr = pDevice->OpenPropertyStore(STGM_READ, &pProperty);
-    if (FAILED(hr))
-    {
-      CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint properties failed.");
-      SAFE_RELEASE(pDevice);
-      goto failed;
-    }
-
-    hr = pProperty->GetValue(PKEY_Device_FriendlyName, &varName);
-    if (FAILED(hr))
-    {
-      CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint device name failed.");
-      SAFE_RELEASE(pDevice);
-      SAFE_RELEASE(pProperty);
-      goto failed;
-    }
-
-    std::string strFriendlyName = localWideToUtf(varName.pwszVal);
-    PropVariantClear(&varName);
-
-    hr = pProperty->GetValue(PKEY_AudioEndpoint_GUID, &varName);
-    if(FAILED(hr))
-    {
-      CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint GUID failed.");
-      SAFE_RELEASE(pDevice);
-      SAFE_RELEASE(pProperty);
-      goto failed;
-    }
-
-    std::string strDevName = localWideToUtf(varName.pwszVal);
-    PropVariantClear(&varName);
-
-    hr = pProperty->GetValue(PKEY_AudioEndpoint_FormFactor, &varName);
-    if (FAILED(hr))
-    {
-      CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint form factor failed.");
-      SAFE_RELEASE(pDevice);
-      SAFE_RELEASE(pProperty);
-      goto failed;
-    }
-    std::string strWinDevType = winEndpoints[(EndpointFormFactor)varName.uiVal].winEndpointType;
-    AEDeviceType aeDeviceType = winEndpoints[(EndpointFormFactor)varName.uiVal].aeDeviceType;
-
-    PropVariantClear(&varName);
-
-    hr = pProperty->GetValue(PKEY_AudioEndpoint_PhysicalSpeakers, &varName);
-    if (FAILED(hr))
-    {
-      CLog::Log(LOGERROR, __FUNCTION__": Retrieval of WASAPI endpoint speaker layout failed.");
-      SAFE_RELEASE(pDevice);
-      SAFE_RELEASE(pProperty);
-      goto failed;
-    }
-    unsigned int uiChannelMask = std::max(varName.uintVal, (unsigned int) KSAUDIO_SPEAKER_STEREO);
-
-    deviceChannels.Reset();
-
-    for (unsigned int c = 0; c < WASAPI_SPEAKER_COUNT; c++)
-    {
-      if (uiChannelMask & WASAPIChannelOrder[c])
-        deviceChannels += AEChannelNames[c];
-    }
-
-    PropVariantClear(&varName);
-
     IAudioClient *pClient;
-    hr = pDevice->Activate(IID_IAudioClient, CLSCTX_ALL, NULL, (void**)&pClient);
+    hr = pDevice->Activate(&pClient);
     if (SUCCEEDED(hr))
     {
       /* Test format DTS-HD */
@@ -686,10 +411,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       wfxex.Format.nBlockAlign          = wfxex.Format.nChannels * (wfxex.Format.wBitsPerSample >> 3);
       wfxex.Format.nAvgBytesPerSec      = wfxex.Format.nSamplesPerSec * wfxex.Format.nBlockAlign;
       hr = pClient->IsFormatSupported(AUDCLNT_SHAREMODE_EXCLUSIVE, &wfxex.Format, NULL);
-      if (SUCCEEDED(hr) || aeDeviceType == AE_DEVTYPE_HDMI)
+      if (SUCCEEDED(hr) || details.eDeviceType == AE_DEVTYPE_HDMI)
       {
         if(FAILED(hr))
-          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_DTSHD), strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_DTSHD), details.strDescription.c_str());
 
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_DTSHD);
         add192 = true;
@@ -698,10 +423,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       /* Test format Dolby TrueHD */
       wfxex.SubFormat                   = KSDATAFORMAT_SUBTYPE_IEC61937_DOLBY_MLP;
       hr = pClient->IsFormatSupported(AUDCLNT_SHAREMODE_EXCLUSIVE, &wfxex.Format, NULL);
-      if (SUCCEEDED(hr) || aeDeviceType == AE_DEVTYPE_HDMI)
+      if (SUCCEEDED(hr) || details.eDeviceType == AE_DEVTYPE_HDMI)
       {
         if(FAILED(hr))
-          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_TRUEHD), strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_TRUEHD), details.strDescription.c_str());
 
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_TRUEHD);
         add192 = true;
@@ -713,10 +438,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       wfxex.Format.nBlockAlign          = wfxex.Format.nChannels * (wfxex.Format.wBitsPerSample >> 3);
       wfxex.Format.nAvgBytesPerSec      = wfxex.Format.nSamplesPerSec * wfxex.Format.nBlockAlign;
       hr = pClient->IsFormatSupported(AUDCLNT_SHAREMODE_EXCLUSIVE, &wfxex.Format, NULL);
-      if (SUCCEEDED(hr) || aeDeviceType == AE_DEVTYPE_HDMI)
+      if (SUCCEEDED(hr) || details.eDeviceType == AE_DEVTYPE_HDMI)
       {
         if(FAILED(hr))
-          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_EAC3), strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_EAC3), details.strDescription.c_str());
 
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_EAC3);
         add192 = true;
@@ -729,10 +454,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       wfxex.Format.nBlockAlign          = wfxex.Format.nChannels * (wfxex.Format.wBitsPerSample >> 3);
       wfxex.Format.nAvgBytesPerSec      = wfxex.Format.nSamplesPerSec * wfxex.Format.nBlockAlign;
       hr = pClient->IsFormatSupported(AUDCLNT_SHAREMODE_EXCLUSIVE, &wfxex.Format, NULL);
-      if (SUCCEEDED(hr) || aeDeviceType == AE_DEVTYPE_HDMI)
+      if (SUCCEEDED(hr) || details.eDeviceType == AE_DEVTYPE_HDMI)
       {
         if(FAILED(hr))
-          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", "STREAM_TYPE_DTS", strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", "STREAM_TYPE_DTS", details.strDescription.c_str());
 
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_DTSHD_CORE);
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_DTS_2048);
@@ -743,10 +468,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       /* Test format Dolby AC3 */
       wfxex.SubFormat                   = KSDATAFORMAT_SUBTYPE_IEC61937_DOLBY_DIGITAL;
       hr = pClient->IsFormatSupported(AUDCLNT_SHAREMODE_EXCLUSIVE, &wfxex.Format, NULL);
-      if (SUCCEEDED(hr) || aeDeviceType == AE_DEVTYPE_HDMI)
+      if (SUCCEEDED(hr) || details.eDeviceType == AE_DEVTYPE_HDMI)
       {
         if(FAILED(hr))
-          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_AC3), strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": stream type \"%s\" on device \"%s\" seems to be not supported.", CAEUtil::StreamTypeToStr(CAEStreamInfo::STREAM_TYPE_AC3), details.strDescription.c_str());
 
         deviceInfo.m_streamTypes.push_back(CAEStreamInfo::STREAM_TYPE_AC3);
       }
@@ -804,7 +529,7 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
         else if (wfxex.Format.nSamplesPerSec == 192000 && add192)
         {
           deviceInfo.m_sampleRates.push_back(WASAPISampleRates[j]);
-          CLog::Log(LOGNOTICE, __FUNCTION__": sample rate 192khz on device \"%s\" seems to be not supported.", strFriendlyName.c_str());
+          CLog::Log(LOGNOTICE, __FUNCTION__": sample rate 192khz on device \"%s\" seems to be not supported.", details.strDescription.c_str());
         }
       }
       pClient->Release();
@@ -814,10 +539,10 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
       CLog::Log(LOGDEBUG, __FUNCTION__": Failed to activate device for passthrough capability testing.");
     }
 
-    deviceInfo.m_deviceName       = strDevName;
-    deviceInfo.m_displayName      = strWinDevType.append(strFriendlyName);
-    deviceInfo.m_displayNameExtra = std::string("WASAPI: ").append(strFriendlyName);
-    deviceInfo.m_deviceType       = aeDeviceType;
+    deviceInfo.m_deviceName       = details.strDeviceId;
+    deviceInfo.m_displayName      = details.strWinDevType.append(details.strDescription);
+    deviceInfo.m_displayNameExtra = std::string("WASAPI: ").append(details.strDescription);
+    deviceInfo.m_deviceType       = details.eDeviceType;
     deviceInfo.m_channels         = deviceChannels;
 
     /* Store the device info */
@@ -828,21 +553,16 @@ void CAESinkWASAPI::EnumerateDevicesEx(AEDeviceInfoList &deviceInfoList, bool fo
 
     deviceInfoList.push_back(deviceInfo);
 
-    if(pDevice->GetId(&pwszID) == S_OK)
+    if(details.bDefault)
     {
-      if(wstrDDID.compare(pwszID) == 0)
-      {
-        deviceInfo.m_deviceName = std::string("default");
-        deviceInfo.m_displayName = std::string("default");
-        deviceInfo.m_displayNameExtra = std::string("");
-        deviceInfo.m_wantsIECPassthrough = true;
-        deviceInfoList.push_back(deviceInfo);
-      }
-      CoTaskMemFree(pwszID);
+      deviceInfo.m_deviceName = std::string("default");
+      deviceInfo.m_displayName = std::string("default");
+      deviceInfo.m_displayNameExtra = std::string("");
+      deviceInfo.m_wantsIECPassthrough = true;
+      deviceInfoList.push_back(deviceInfo);
     }
 
     SAFE_RELEASE(pDevice);
-    SAFE_RELEASE(pProperty);
   }
   return;
 
@@ -850,94 +570,14 @@ failed:
 
   if (FAILED(hr))
     CLog::Log(LOGERROR, __FUNCTION__": Failed to enumerate WASAPI endpoint devices (%s).", WASAPIErrToStr(hr));
-
-  SAFE_RELEASE(pEnumDevices);
-  SAFE_RELEASE(pEnumerator);
 }
 
 //Private utility functions////////////////////////////////////////////////////
 
-void CAESinkWASAPI::BuildWaveFormatExtensible(AEAudioFormat &format, WAVEFORMATEXTENSIBLE &wfxex)
-{
-  wfxex.Format.wFormatTag        = WAVE_FORMAT_EXTENSIBLE;
-  wfxex.Format.cbSize            = sizeof(WAVEFORMATEXTENSIBLE)-sizeof(WAVEFORMATEX);
-
-
-  if (format.m_dataFormat != AE_FMT_RAW) // PCM data
-  {
-    wfxex.dwChannelMask          = SpeakerMaskFromAEChannels(format.m_channelLayout);
-    wfxex.Format.nChannels       = (WORD)format.m_channelLayout.Count();
-    wfxex.Format.nSamplesPerSec  = format.m_sampleRate;
-    wfxex.Format.wBitsPerSample  = CAEUtil::DataFormatToBits((AEDataFormat) format.m_dataFormat);
-    wfxex.SubFormat              = format.m_dataFormat < AE_FMT_FLOAT ? KSDATAFORMAT_SUBTYPE_PCM : KSDATAFORMAT_SUBTYPE_IEEE_FLOAT;
-  }
-  else //Raw bitstream
-  {
-    wfxex.Format.wFormatTag = WAVE_FORMAT_EXTENSIBLE;
-    if (format.m_dataFormat == AE_FMT_RAW &&
-        ((format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_AC3) ||
-         (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_EAC3) ||
-         (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_DTSHD_CORE) ||
-         (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_DTS_2048) ||
-         (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_DTS_1024) ||
-         (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_DTS_512)))
-    {
-      if (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_EAC3)
-        wfxex.SubFormat = KSDATAFORMAT_SUBTYPE_IEC61937_DOLBY_DIGITAL_PLUS;
-      else
-        wfxex.SubFormat = KSDATAFORMAT_SUBTYPE_IEC61937_DOLBY_DIGITAL;
-      wfxex.dwChannelMask               = bool (format.m_channelLayout.Count() == 2) ? KSAUDIO_SPEAKER_STEREO : KSAUDIO_SPEAKER_5POINT1;
-      wfxex.Format.wBitsPerSample       = 16;
-      wfxex.Samples.wValidBitsPerSample = 16;
-      wfxex.Format.nChannels            = (WORD)format.m_channelLayout.Count();
-      wfxex.Format.nSamplesPerSec       = format.m_sampleRate;
-      if (format.m_streamInfo.m_sampleRate == 0)
-      CLog::Log(LOGERROR, "Invalid sample rate supplied for RAW format");
-    }
-    else if (format.m_dataFormat == AE_FMT_RAW &&
-             ((format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_DTSHD) ||
-              (format.m_streamInfo.m_type == CAEStreamInfo::STREAM_TYPE_TRUEHD)))
-    {
-      // IEC 61937 transmissions over HDMI       
-      wfxex.Format.nSamplesPerSec       = 192000L;
-      wfxex.Format.wBitsPerSample       = 16;
-      wfxex.Samples.wValidBitsPerSample = 16;
-      wfxex.dwChannelMask               = KSAUDIO_SPEAKER_7POINT1_SURROUND;
-
-      switch (format.m_streamInfo.m_type)
-      {
-        case CAEStreamInfo::STREAM_TYPE_TRUEHD:
-          wfxex.SubFormat             = KSDATAFORMAT_SUBTYPE_IEC61937_DOLBY_MLP;
-          wfxex.Format.nChannels      = 8; // Four IEC 60958 Lines.
-          wfxex.dwChannelMask         = KSAUDIO_SPEAKER_7POINT1_SURROUND;
-          break;
-        case CAEStreamInfo::STREAM_TYPE_DTSHD:
-          wfxex.SubFormat             = KSDATAFORMAT_SUBTYPE_IEC61937_DTS_HD;
-          wfxex.Format.nChannels      = 8; // Four IEC 60958 Lines.
-          wfxex.dwChannelMask         = KSAUDIO_SPEAKER_7POINT1_SURROUND;
-          break;
-      }
-
-      if (format.m_channelLayout.Count() == 8)
-        wfxex.dwChannelMask = KSAUDIO_SPEAKER_7POINT1_SURROUND;
-      else
-        wfxex.dwChannelMask = KSAUDIO_SPEAKER_5POINT1;
-    }
-  }
-
-  if (format.m_dataFormat == AE_FMT_S24NE4MSB)
-    wfxex.Samples.wValidBitsPerSample = 24;
-  else
-    wfxex.Samples.wValidBitsPerSample = wfxex.Format.wBitsPerSample;
-
-  wfxex.Format.nBlockAlign          = wfxex.Format.nChannels * (wfxex.Format.wBitsPerSample >> 3);
-  wfxex.Format.nAvgBytesPerSec      = wfxex.Format.nSamplesPerSec * wfxex.Format.nBlockAlign;
-}
-
 void CAESinkWASAPI::BuildWaveFormatExtensibleIEC61397(AEAudioFormat &format, WAVEFORMATEXTENSIBLE_IEC61937 &wfxex)
 {
   /* Fill the common structure */
-  BuildWaveFormatExtensible(format, wfxex.FormatExt);
+  CAESinkFactoryWin::BuildWaveFormatExtensible(format, wfxex.FormatExt);
 
   /* Code below kept for future use - preferred for later Windows versions */
   /* but can cause problems on older Windows versions and drivers          */
@@ -956,14 +596,14 @@ bool CAESinkWASAPI::InitializeExclusive(AEAudioFormat &format)
   WAVEFORMATEXTENSIBLE &wfxex = wfxex_iec61937.FormatExt;
 
   if (format.m_dataFormat <= AE_FMT_FLOAT)
-    BuildWaveFormatExtensible(format, wfxex);
+    CAESinkFactoryWin::BuildWaveFormatExtensible(format, wfxex);
   else if (format.m_dataFormat == AE_FMT_RAW)
     BuildWaveFormatExtensibleIEC61397(format, wfxex_iec61937);
   else
   {
     // planar formats are currently not supported by this sink
     format.m_dataFormat = AE_FMT_FLOAT;
-    BuildWaveFormatExtensible(format, wfxex);
+    CAESinkFactoryWin::BuildWaveFormatExtensible(format, wfxex);
   }
 
   /* Test for incomplete format and provide defaults */
@@ -1019,7 +659,7 @@ bool CAESinkWASAPI::InitializeExclusive(AEAudioFormat &format)
     }
     else if (layout >= 0)
     {
-      wfxex.dwChannelMask = ChLayoutToChMask(layoutsList[layout], &noOfCh);
+      wfxex.dwChannelMask = CAESinkFactoryWin::ChLayoutToChMask(layoutsList[layout], &noOfCh);
       wfxex.Format.nChannels = noOfCh;
       if (noOfCh < requestedChannels)
         continue;
@@ -1080,7 +720,7 @@ bool CAESinkWASAPI::InitializeExclusive(AEAudioFormat &format)
 
 initialize:
 
-  AEChannelsFromSpeakerMask(wfxex.dwChannelMask);
+  CAESinkFactoryWin::AEChannelsFromSpeakerMask(m_channelLayout, wfxex.dwChannelMask);
   format.m_channelLayout = m_channelLayout;
 
   /* When the stream is raw, the values in the format structure are set to the link    */
@@ -1144,7 +784,7 @@ initialize:
     SAFE_RELEASE(m_pAudioClient);
 
     /* Create a new audio client */
-    hr = m_pDevice->Activate(IID_IAudioClient, CLSCTX_ALL, NULL, (void**)&m_pAudioClient);
+    hr = m_pDevice->Activate(&m_pAudioClient);
     if (FAILED(hr))
     {
       CLog::Log(LOGERROR, __FUNCTION__": Device Activation Failed : %s", WASAPIErrToStr(hr));
@@ -1196,67 +836,6 @@ initialize:
   return true;
 }
 
-void CAESinkWASAPI::AEChannelsFromSpeakerMask(DWORD speakers)
-{
-  m_channelLayout.Reset();
-
-  for (int i = 0; i < WASAPI_SPEAKER_COUNT; i++)
-  {
-    if (speakers & WASAPIChannelOrder[i])
-      m_channelLayout += AEChannelNames[i];
-  }
-}
-
-DWORD CAESinkWASAPI::SpeakerMaskFromAEChannels(const CAEChannelInfo &channels)
-{
-  DWORD mask = 0;
-
-  for (unsigned int i = 0; i < channels.Count(); i++)
-  {
-    for (unsigned int j = 0; j < WASAPI_SPEAKER_COUNT; j++)
-      if (channels[i] == AEChannelNames[j])
-        mask |= WASAPIChannelOrder[j];
-  }
-  return mask;
-}
-
-const char *CAESinkWASAPI::WASAPIErrToStr(HRESULT err)
-{
-  switch(err)
-  {
-    ERRTOSTR(AUDCLNT_E_NOT_INITIALIZED);
-    ERRTOSTR(AUDCLNT_E_ALREADY_INITIALIZED);
-    ERRTOSTR(AUDCLNT_E_WRONG_ENDPOINT_TYPE);
-    ERRTOSTR(AUDCLNT_E_DEVICE_INVALIDATED);
-    ERRTOSTR(AUDCLNT_E_NOT_STOPPED);
-    ERRTOSTR(AUDCLNT_E_BUFFER_TOO_LARGE);
-    ERRTOSTR(AUDCLNT_E_OUT_OF_ORDER);
-    ERRTOSTR(AUDCLNT_E_UNSUPPORTED_FORMAT);
-    ERRTOSTR(AUDCLNT_E_INVALID_SIZE);
-    ERRTOSTR(AUDCLNT_E_DEVICE_IN_USE);
-    ERRTOSTR(AUDCLNT_E_BUFFER_OPERATION_PENDING);
-    ERRTOSTR(AUDCLNT_E_THREAD_NOT_REGISTERED);
-    ERRTOSTR(AUDCLNT_E_EXCLUSIVE_MODE_NOT_ALLOWED);
-    ERRTOSTR(AUDCLNT_E_ENDPOINT_CREATE_FAILED);
-    ERRTOSTR(AUDCLNT_E_SERVICE_NOT_RUNNING);
-    ERRTOSTR(AUDCLNT_E_EVENTHANDLE_NOT_EXPECTED);
-    ERRTOSTR(AUDCLNT_E_EXCLUSIVE_MODE_ONLY);
-    ERRTOSTR(AUDCLNT_E_BUFDURATION_PERIOD_NOT_EQUAL);
-    ERRTOSTR(AUDCLNT_E_EVENTHANDLE_NOT_SET);
-    ERRTOSTR(AUDCLNT_E_INCORRECT_BUFFER_SIZE);
-    ERRTOSTR(AUDCLNT_E_BUFFER_SIZE_ERROR);
-    ERRTOSTR(AUDCLNT_E_CPUUSAGE_EXCEEDED);
-    ERRTOSTR(AUDCLNT_E_BUFFER_ERROR);
-    ERRTOSTR(AUDCLNT_E_BUFFER_SIZE_NOT_ALIGNED);
-    ERRTOSTR(AUDCLNT_E_INVALID_DEVICE_PERIOD);
-    ERRTOSTR(E_POINTER);
-    ERRTOSTR(E_INVALIDARG);
-    ERRTOSTR(E_OUTOFMEMORY);
-    default: break;
-  }
-  return NULL;
-}
-
 void CAESinkWASAPI::Drain()
 {
   if(!m_pAudioClient)
@@ -1285,20 +864,5 @@ void CAESinkWASAPI::Drain()
 
 bool CAESinkWASAPI::IsUSBDevice()
 {
-  IPropertyStore *pProperty = NULL;
-  PROPVARIANT varName;
-  PropVariantInit(&varName);
-  bool ret = false;
-
-  HRESULT hr = m_pDevice->OpenPropertyStore(STGM_READ, &pProperty);
-  if (!SUCCEEDED(hr))
-    return ret;
-  hr = pProperty->GetValue(PKEY_Device_EnumeratorName, &varName);
-
-  std::string str = localWideToUtf(varName.pwszVal);
-  StringUtils::ToUpper(str);
-  ret = (str == "USB");
-  PropVariantClear(&varName);
-  SAFE_RELEASE(pProperty);
-  return ret;
+  return m_pDevice && m_pDevice->IsUSBDevice();
 }
